Repository Impl with test case.

+++ lib/features/number_trivia/domain/entities/number_trivia.dart
@@ -1,12 +1,10 @@
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 
-class NumberTrivia extends Equatable{
+class NumberTrivia extends Equatable {
   final String text;
-  final int number;
+  final double number;
 
-  NumberTrivia({
-    @required this.text, 
-    @required this.number}) : super([ number, text]);
-  
+  NumberTrivia({@required this.text, @required this.number})
+      : super([number, text]);
 }
